add release_utils module
