+++ scripts/rock_tagging.py
@@ -9,30 +9,28 @@ Automates tag and GitHub release creation for TheRock plus every tracked
 submodule. Authentication is done via SSH.
 
 High-level workflow:
-1. Reuse (or populate) a cached clone under a configurable directory
-    (default: `/tmp/rock-tagging-cache`, overridable via `--cache-dir`),
-    fetch the latest refs, and hard-reset to the user-specified commit.
-2. Update submodules via `fetch_sources.py` when available (fallback to
-    `git submodule update`) and build a plan by combining `.gitmodules`
-    metadata with `git submodule status` output. Repos listed in
-    `--exclude-list` and repos outside the ROCm GitHub org are skipped.
-3. For each component (inside a single loop):
-    a. Configure an SSH `rocm-github` remote.
-    b. Create an annotated tag (`therock-<version>`) at the recorded
-       commit, skipping components where the tag already exists.
-    c. For mono-repos (`rocm-libraries`, `rocm-systems`), generate
-       tarballs for the `projects/` and `shared/` directories (tarballs
-       are created even in dry-run mode).
-    d. When not in dry-run mode, push the tag and invoke
-       `gh release create` with the appropriate notes and tarball assets.
+1. Verify GitHub push/admin permissions for all repos (via GitHub API, no
+   clone required).
+2. Reuse (or populate) a cached clone under a configurable directory
+   (default: `/tmp/rock-tagging-cache`, overridable via `--cache-dir`),
+   fetch the latest refs, and hard-reset to the user-specified commit.
+3. Update submodules via `fetch_sources.py` when available (fallback to
+   `git submodule update`) and build a plan by combining `.gitmodules`
+   metadata with `git submodule status` output. Repos listed in
+   `--exclude-list` and repos outside the ROCm GitHub org are skipped.
+4. For each component:
+   a. Configure an SSH `rocm-github` remote.
+   b. Create an annotated tag (`therock-<version>`) at the recorded
+      commit, skipping components where the tag already exists.
+   c. For mono-repos (`rocm-libraries`, `rocm-systems`), generate
+      tarballs for the `projects/` and `shared/` directories.
+   d. When not in dry-run mode, push the tag and invoke
+      `gh release create` with the appropriate notes and tarball assets.
 
-Use `--force-clone` to delete and reclone when the cache directory exists
-but is not a valid git repo. All actions are logged for traceability, and
-dry-run mode (the default) lets you preview the plan without touching
-remotes.
+Dry-run mode (the default) lets you preview the plan without touching remotes.
 
 Usage:
-        python rock-tagging.py \\
+        python rock_tagging.py \\
                 --branch-name <release-branch> \\
                 --release-version <version> \\
                 --commitid <rock-commit-sha> \\
@@ -52,418 +50,57 @@ Options:
 """
 import argparse
 import logging
-import shlex
-import shutil
 import subprocess
 import sys
 import tarfile
 import tempfile
-from dataclasses import dataclass
 from pathlib import Path
 from pprint import pformat
 
+from release_utils import (
+    RockBase,
+    RepoInfo,
+    TIMEOUT_LONG,
+    check_permissions,
+    fetch_lightweight_plan,
+    get_gh_token,
+)
 
-@dataclass
-class RepoInfo:
-    """Information about a repository to tag."""
 
-    url: str
-    commit: str
-    path: Path
-
-
-class RockTagging:
+class RockTagging(RockBase):
     """Automates tagging and release uploading for TheRock."""
+
+    _cache_dir_name = "rock-tagging-cache"
 
     MONO_REPOS = frozenset({"rocm-libraries", "rocm-systems"})
 
-    # Timeouts (seconds) for subprocess calls.
-    TIMEOUT_LONG = 1800   # clone, fetch, submodule update
-    TIMEOUT_SHORT = 60    # tag, push, git config reads
-
     def __init__(self, cli_args: argparse.Namespace) -> None:
-        self.release_branch: str = cli_args.branch_name
+        super().__init__(cli_args)
         self.release_version: str = cli_args.release_version
-        self.dry_run: bool = cli_args.dry_run
-        self.commitid: str = cli_args.commitid
-        self.exclude_list: set[str] = set(cli_args.exclude_list or [])
-        self.force_clone: bool = cli_args.force_clone
-        self.cache_dir: Path | None = (
-            Path(cli_args.cache_dir) if cli_args.cache_dir else None
-        )
-        self.rock_url: str = "https://github.com/ROCm/TheRock.git"
-        self.cache_root: Path | None = None
-
-        self._logger = logging.getLogger("rock_tagging")
 
         self.log("Authentication Mode: SSH")
         self.log(f"Dry run mode = {self.dry_run}")
         if self.exclude_list:
             self.log(f"Exclude list: {self.exclude_list}")
 
-    def log(self, msg: str) -> None:
-        """Log an info-level message."""
-        self._logger.info(msg)
-
-    def run_command(
-        self,
-        args: list[str | Path],
-        cwd: Path,
-        *,
-        input_data: bytes | None = None,
-        stream: bool = False,
-        timeout: int | None = TIMEOUT_SHORT,
-    ) -> None:
-        """Execute a subprocess command, raising CalledProcessError on failure.
-
-        Args:
-            args:       Command and arguments to execute.
-            cwd:        Working directory for the command.
-            input_data: Optional bytes piped to stdin.
-            stream:     If True, print stdout/stderr line-by-line as it arrives
-                        (useful for long-running operations like clone/fetch).
-                        If False, buffer output and log after completion.
-            timeout:    Seconds before the process is killed (default: TIMEOUT_SHORT).
-                        Pass TIMEOUT_LONG for clone/fetch/submodule operations.
-        """
-        cmd = args if isinstance(args, list) else [args]
-        self.log(f"++ Exec [{cwd}]$ {shlex.join(map(str, cmd))}")
-        sys.stdout.flush()
-
-        if stream:
-            process = subprocess.Popen(
-                cmd,
-                cwd=str(cwd),
-                stdout=subprocess.PIPE,
-                stderr=subprocess.STDOUT,
-                text=True,
-            )
-
-            try:
-                for line in process.stdout:
-                    self.log(line.rstrip())
-                ret = process.wait(timeout=timeout)
-            except subprocess.TimeoutExpired:
-                process.kill()
-                process.wait()
-                raise subprocess.TimeoutExpired(cmd, timeout)
-            if ret != 0:
-                raise subprocess.CalledProcessError(ret, cmd)
-
-            return
-
-        try:
-            result = subprocess.run(
-                cmd,
-                cwd=str(cwd),
-                shell=False,
-                input=input_data,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                check=True,
-                stdin=None if input_data else subprocess.DEVNULL,
-                text=False,
-                timeout=timeout,
-            )
-
-            if result.stdout:
-                self.log(
-                    result.stdout
-                    if isinstance(result.stdout, str)
-                    else result.stdout.decode(errors="ignore")
-                )
-            if result.stderr:
-                self.log(
-                    result.stderr
-                    if isinstance(result.stderr, str)
-                    else result.stderr.decode(errors="ignore")
-                )
-
-        except subprocess.CalledProcessError as exc:
-            self.log(
-                (exc.stdout or b"").decode(errors="ignore")
-                if isinstance(exc.stdout, bytes)
-                else (exc.stdout or "")
-            )
-            self.log(
-                (exc.stderr or b"").decode(errors="ignore")
-                if isinstance(exc.stderr, bytes)
-                else (exc.stderr or "")
-            )
-            raise
-
-    def run_command_output(
-        self,
-        args: list[str | Path],
-        cwd: Path,
-        timeout: int | None = TIMEOUT_SHORT,
-    ) -> str:
-        """Run a command and return its stripped stdout as a string.
-
-        Raises CalledProcessError on non-zero exit.
-        """
-        cmd = args if isinstance(args, list) else [args]
-        self.log(f"++ Exec [{cwd}]$ {shlex.join(map(str, cmd))}")
-
-        result = subprocess.run(
-            cmd,
-            cwd=str(cwd),
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-            check=True,
-            stdin=subprocess.DEVNULL,
-            timeout=timeout,
-        )
-        return result.stdout.strip()
-
-    def convert_to_ssh(self, url: str) -> str:
-        """Convert https://github.com/X/Y.git to git@github.com:X/Y.git."""
-        if url.startswith("https://github.com/"):
-            path = url.replace("https://github.com/", "")
-            return f"git@github.com:{path}"
-        return url
-
-    def get_submodule_url_map(self, repo_dir: Path) -> dict[str, str]:
-        """Return mapping of submodule working-tree paths to remote URLs."""
-        gitmodules_path = repo_dir / ".gitmodules"
-        if not gitmodules_path.exists():
-            return {}
-
-        try:
-            path_entries = self.run_command_output(
-                [
-                    "git",
-                    "config",
-                    "--file",
-                    str(gitmodules_path),
-                    "--get-regexp",
-                    r"submodule\..*\.path",
-                ],
-                cwd=repo_dir,
-            )
-        except subprocess.CalledProcessError:
-            return {}
-
-        url_map: dict[str, str] = {}
-        for line in path_entries.splitlines():
-            # Each line looks like:
-            #   "submodule.external/hipcc.path external/hipcc"
-            parts = line.strip().split(None, 1)
-            if len(parts) != 2:
-                continue
-            key, path_value = parts
-            section = key.rsplit(".", 1)[0]
-            try:
-                url = self.run_command_output(
-                    [
-                        "git",
-                        "config",
-                        "--file",
-                        str(gitmodules_path),
-                        "--get",
-                        f"{section}.url",
-                    ],
-                    cwd=repo_dir,
-                )
-            except subprocess.CalledProcessError:
-                self.log(f"No URL entry for {section}; skipping")
-                continue
-            url_map[path_value.strip()] = url
-
-        return url_map
-
-    def build_plan(self) -> dict[str, RepoInfo]:
-        """Build the tagging execution plan.
-
-        1. Clone (or reuse cached clone of) TheRock.
-        2. Check out and hard-reset to ``self.commitid``.
-        3. Populate submodules via ``fetch_sources.py`` (or ``git submodule update``).
-        4. Read ``git submodule status`` and ``.gitmodules`` to collect each
-           submodule's commit SHA, remote URL, and local path.
-        5. Return a dict keyed by repo name, including TheRock itself.
-        """
-        cache_root = (
-            self.cache_dir
-            or Path(tempfile.gettempdir()) / "rock-tagging-cache"
-        )
-        cache_root.mkdir(parents=True, exist_ok=True)
-        clone_dir = cache_root / "TheRock"
-        self.cache_root = cache_root
-
-        needs_clone = not clone_dir.exists()
-        if not needs_clone and not (clone_dir / ".git").exists():
-            if not self.force_clone:
-                raise RuntimeError(
-                    f"Cache directory {clone_dir} exists but is not a git "
-                    "repo. Use --force-clone to delete it and reclone."
-                )
-            self.log(
-                f"Cache directory {clone_dir} is not a git repo; "
-                "removing before reclone (--force-clone)"
-            )
-            shutil.rmtree(clone_dir)
-            needs_clone = True
-
-        if needs_clone:
-            self.log(
-                f"Cloning TheRock repo from {self.rock_url} into {clone_dir}"
-            )
-            self.run_command(
-                ["git", "clone", str(self.rock_url), str(clone_dir)],
-                cwd=cache_root,
-                stream=True,
-                timeout=self.TIMEOUT_LONG,
-            )
-        else:
-            self.log(f"Reusing existing TheRock repo at {clone_dir}")
-
-            try:
-                remote_url = self.run_command_output(
-                    ["git", "remote", "get-url", "origin"],
-                    cwd=clone_dir,
-                )
-                if "TheRock" not in remote_url:
-                    raise RuntimeError(
-                        f"Existing repo at {clone_dir} does not look like "
-                        f"TheRock (origin={remote_url})"
-                    )
-            except subprocess.CalledProcessError as exc:
-                raise RuntimeError(
-                    f"Failed to inspect existing repo at {clone_dir}: {exc}"
-                ) from exc
-
-            self.log("Fetching latest changes for existing TheRock clone...")
-            self.run_command(
-                [
-                    "git",
-                    "fetch",
-                    "origin",
-                    "--prune",
-                    "--recurse-submodules=on-demand",
-                ],
-                cwd=clone_dir,
-                stream=True,
-                timeout=self.TIMEOUT_LONG,
-            )
-
-        fetch_script = clone_dir / "build_tools" / "fetch_sources.py"
+    def _checkout_and_update_submodules(self, clone_dir: Path) -> None:
+        """Fetch the release branch then delegate to the base checkout logic."""
         rock_commit = self.commitid
-
         self.log(
             f"Fetching release branch '{self.release_branch}' to ensure "
             f"commit {rock_commit} is reachable..."
         )
         self.run_command(
             [
-                "git",
-                "fetch",
-                "origin",
-                f"refs/heads/{self.release_branch}:refs/remotes/origin/{self.release_branch}",
+                "git", "fetch", "origin",
+                f"refs/heads/{self.release_branch}:"
+                f"refs/remotes/origin/{self.release_branch}",
             ],
             cwd=clone_dir,
             stream=True,
-            timeout=self.TIMEOUT_LONG,
+            timeout=TIMEOUT_LONG,
         )
-
-        self.log(f"Checking out TheRock at commit {rock_commit}")
-        self.run_command(["git", "checkout", rock_commit], cwd=clone_dir)
-        self.run_command(
-            ["git", "reset", "--hard", rock_commit], cwd=clone_dir
-        )
-
-        if fetch_script.exists():
-            self.log(
-                "Updating submodules via fetch_sources.py "
-                "(jobs=10, no patches)..."
-            )
-            self.run_command(
-                [
-                    "python3",
-                    str(fetch_script),
-                    "--jobs",
-                    "10",
-                    "--no-apply-patches",
-                ],
-                cwd=clone_dir,
-                stream=True,
-                timeout=self.TIMEOUT_LONG,
-            )
-        else:
-            self.log(
-                "fetch_sources.py not found; "
-                "falling back to git submodule update"
-            )
-            self.run_command(
-                ["git", "submodule", "update", "--init", "--recursive"],
-                cwd=clone_dir,
-                stream=True,
-                timeout=self.TIMEOUT_LONG,
-            )
-
-        self.log("Reading submodule status...")
-        try:
-            status_output = self.run_command_output(
-                ["git", "submodule", "status"],
-                cwd=clone_dir,
-            )
-            lines = status_output.split("\n") if status_output else []
-        except subprocess.CalledProcessError as exc:
-            raise RuntimeError(
-                f"Failed to read submodule status: {exc}"
-            ) from exc
-
-        url_map = self.get_submodule_url_map(clone_dir)
-
-        plan: dict[str, RepoInfo] = {}
-
-        for line in lines:
-            if not line:
-                continue
-
-            parts = line.split()
-            if len(parts) < 2:
-                continue
-            sha = parts[0].lstrip("-+")
-            path = parts[1]
-
-            repo_name = Path(path).name
-            repo_url = url_map.get(path)
-
-            if not repo_url:
-                self.log(
-                    f"No URL found for submodule {path} in .gitmodules"
-                )
-                continue
-
-            if repo_name in self.exclude_list:
-                self.log(f"Skipping {repo_name} (in exclude list)")
-                continue
-
-            url_lower = repo_url.lower()
-            if (
-                "github.com/rocm/" not in url_lower
-                and "github.com:rocm/" not in url_lower
-            ):
-                self.log(
-                    f"Skipping {repo_name} "
-                    f"(not a ROCm org repo: {repo_url})"
-                )
-                continue
-
-            plan[repo_name] = RepoInfo(
-                url=repo_url,
-                commit=sha,
-                path=clone_dir / path,
-            )
-
-        plan["TheRock"] = RepoInfo(
-            url=self.rock_url,
-            commit=rock_commit,
-            path=clone_dir,
-        )
-
-        return plan
+        super()._checkout_and_update_submodules(clone_dir)
 
     def _create_tarballs(
         self,
@@ -472,7 +109,7 @@ class RockTagging:
         tarball_paths: list[Path],
         label: str,
     ) -> None:
-        """Create per-subdirectory tarballs from *source_dir*."""
+        """Create per-subdirectory tarballs from source_dir."""
         if not source_dir.is_dir():
             self.log(f"Source directory not found for {label}: {source_dir}")
             return
@@ -481,39 +118,16 @@ class RockTagging:
         for entry in sorted(source_dir.iterdir()):
             if entry.name.startswith(".") or not entry.is_dir():
                 continue
-
             tarball_path = root_dir / f"{entry.name}.tar.gz"
             if tarball_path in tarball_paths:
                 continue
-
             with tarfile.open(tarball_path, "w:gz") as tf:
                 tf.add(str(entry), arcname=entry.name)
             tarball_paths.append(tarball_path)
             self.log(f"Tarball created: {tarball_path}")
 
-    def _setup_remote(self, url: str, repo_dir: Path) -> None:
-        """Add or update the rocm-github remote for a repo."""
-        remote_url = self.convert_to_ssh(url)
-        try:
-            self.run_command(
-                ["git", "remote", "set-url", "rocm-github", remote_url],
-                cwd=repo_dir,
-            )
-        except subprocess.CalledProcessError:
-            self.run_command(
-                ["git", "remote", "add", "rocm-github", remote_url],
-                cwd=repo_dir,
-            )
-
     def execute_plan(self, plan: dict[str, RepoInfo]) -> None:
-        """Execute the tagging plan for every repo in *plan*.
-
-        For each repo:
-        1. Configure the SSH ``rocm-github`` remote.
-        2. Create an annotated tag at the recorded commit.
-        3. For mono-repos, generate tarballs.
-        4. Push the tag and create a GitHub release (skipped in dry-run mode).
-        """
+        """Create and push tags, and publish GitHub releases, for every repo."""
         successful_components: dict[str, RepoInfo] = {}
         failed_components: dict[str, str] = {}
         work_dir = self.cache_root or Path(tempfile.gettempdir())
@@ -529,7 +143,6 @@ class RockTagging:
                 failed_components[comp] = f"Remote setup failed: {exc}"
                 continue
 
-            # Skip if tag already exists locally
             tag_exists = subprocess.run(
                 ["git", "rev-parse", "-q", "--verify", tag_name],
                 cwd=info.path,
@@ -538,41 +151,27 @@ class RockTagging:
             ).returncode == 0
 
             if tag_exists:
-                self.log(
-                    f"Tag {tag_name} already exists for {comp}; "
-                    "skipping creation"
-                )
+                self.log(f"Tag {tag_name} already exists for {comp}; skipping creation")
                 successful_components[comp] = info
                 continue
 
             try:
                 self.run_command(
                     [
-                        "git",
-                        "tag",
-                        "-a",
-                        tag_name,
-                        info.commit,
-                        "-m",
-                        f"therock release v{self.release_version}",
+                        "git", "tag", "-a", tag_name, info.commit,
+                        "-m", f"therock release v{self.release_version}",
                     ],
                     cwd=info.path,
                 )
                 if not self.dry_run:
                     self.run_command(
-                        [
-                            "git",
-                            "push",
-                            "rocm-github",
-                            f"{tag_name}:refs/tags/{tag_name}",
-                        ],
+                        ["git", "push", "rocm-github", f"{tag_name}:refs/tags/{tag_name}"],
                         cwd=info.path,
                     )
             except (subprocess.CalledProcessError, subprocess.TimeoutExpired) as exc:
                 failed_components[comp] = f"Tag failed: {exc}"
                 continue
 
-            # Tarballs only for mono-repos
             tarballs: list[Path] = []
             if comp in self.MONO_REPOS:
                 self._create_tarballs(
@@ -586,21 +185,18 @@ class RockTagging:
                 self.log(f"[DRY RUN] Would create release with: {tarballs}")
             else:
                 try:
-                    release_cmd = [
-                        "gh",
-                        "release",
-                        "create",
-                        tag_name,
-                        "--notes",
-                        f"therock release v{self.release_version}",
-                        *[str(p) for p in tarballs],
-                    ]
-                    self.run_command(release_cmd, cwd=info.path)
-                except (subprocess.CalledProcessError, subprocess.TimeoutExpired) as exc:
-                    failed_components[comp] = (
-                        f"Release creation failed: {exc}"
+                    self.run_command(
+                        [
+                            "gh", "release", "create", tag_name,
+                            "--notes", f"therock release v{self.release_version}",
+                            *[str(p) for p in tarballs],
+                        ],
+                        cwd=info.path,
                     )
+                except (subprocess.CalledProcessError, subprocess.TimeoutExpired) as exc:
+                    failed_components[comp] = f"Release creation failed: {exc}"
                     continue
+
             successful_components[comp] = info
 
         self.log(
@@ -613,34 +209,29 @@ class RockTagging:
             self.log(f"Failed components: {pformat(failed_components)}")
 
     def run(self) -> None:
-        """Build the execution plan and execute it."""
+        """Check permissions, build the plan, and execute it."""
+        token = get_gh_token()
+        repo_map = fetch_lightweight_plan(token, self.commitid, self.exclude_list)
+        check_permissions(token, repo_map, self._logger, action="tags")
+
         plan = self.build_plan()
         self.log(f"Execution plan: {pformat(plan)}")
         self.execute_plan(plan)
 
 
 def main(argv: list[str]) -> int:
-    """Parse arguments and run the tagging automation."""
-    parser = argparse.ArgumentParser(
-        description="Rock Tagging Automation Tool",
-    )
+    parser = argparse.ArgumentParser(description="Rock Tagging Automation Tool")
     parser.add_argument(
-        "-B",
-        "--branch-name",
-        required=True,
+        "-B", "--branch-name", required=True,
         help="Name of the release branch",
     )
     parser.add_argument(
-        "-V",
-        "--release-version",
-        required=True,
+        "-V", "--release-version", required=True,
         help="Release version string (used for tag names)",
     )
     parser.add_argument(
-        "-C",
-        "--commitid",
-        required=True,
-        help="Commit ID of TheRock to tag from",
+        "-C", "--commitid", required=True,
+        help="Commit SHA of TheRock to tag from",
     )
     parser.add_argument(
         "--dry-run",
@@ -649,29 +240,20 @@ def main(argv: list[str]) -> int:
         help="Log actions without pushing to remotes (default: enabled)",
     )
     parser.add_argument(
-        "--exclude-list",
-        nargs="*",
-        default=[],
-        help="List of submodule repo names to exclude from tagging",
+        "--exclude-list", nargs="*", default=[],
+        help="Submodule repo names to exclude from tagging",
     )
     parser.add_argument(
-        "--force-clone",
-        action="store_true",
-        default=False,
-        help="Delete and reclone if cache directory exists but is not a "
-        "valid git repo",
+        "--force-clone", action="store_true", default=False,
+        help="Delete and reclone if cache dir exists but is not a valid git repo",
     )
     parser.add_argument(
-        "--cache-dir",
-        default=None,
-        help="Directory to cache the TheRock clone "
-        "(default: /tmp/rock-tagging-cache)",
+        "--cache-dir", default=None,
+        help="Directory to cache the TheRock clone (default: /tmp/rock-tagging-cache)",
     )
     args = parser.parse_args(argv)
 
-    logging.basicConfig(
-        level=logging.INFO, format="[%(levelname)s] %(message)s"
-    )
+    logging.basicConfig(level=logging.INFO, format="[%(levelname)s] %(message)s")
 
     try:
         RockTagging(args).run()
